requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.7
Django==3.1.7
django-simple-history==2.12.0
django-tempus-dominus==5.1.2.13
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
